Support mapping bindings.
Also simplified how composite bindings are referenced. Instead of having
to use an explicit Key() for lists or dictionaries, you can use a list
or dictionary with one element.

For example, a mapping of strings to plugins:

    {str: Plugin}

Or a sequence of plugins:

    [Plugin]

Pass these forms to get(), @extends() and multibind().