Stop supporting assisted injection of callables
This isn't a feature that's used too widely from what I know and it's
rather easy to emulate it if needed. I'm removing it because I plan to
make AssistedBuilder and ProviderOf generic classes[1] so that
AssistedBuilder[T] and ProviderOf[T] can be used in type hints. In order
to do that we need to make sure T is a type, not a callable.

[1] https://docs.python.org/3/library/typing.html#user-defined-generic-types