Fix scope and parent/child injector interaction
The issue was that even though some bindings may have been local to
a child Injector only the root Injector actually had the scope objects
which were shared among all Injectors in a hierarchy. Due to that when
an object was stored in a scope it was being stored in the single (per
hierarchy) shared scope.

This patch removes the arbitrary limitation and all Injectors have
scopes bound locally now.

It's worth noting that while the Scope interface doesn't change here
this patch still breaks backwards compatibility - custom scopes need to
be provided to all Injector instantiations, not only the root one.

This may introduce some subtle issues but I can't think of any right
now.