Change the way scope binding/creation works
Previously Injector had knowledge about a predefined set of default
scopes which meant that when one had a custom scope they had to bind it
explicitly (Binder.bind_scope).

A fix to unwanted sharing objects between parent and children Injectors
complicated the situation because if one wanted to use a custom scope
one had to bind it explicitly with all Injectors in a hierarchy after
the change was made.

This change removes the necessity for the bind_scope step - Injector
will dynamically bind scopes as it encounters references to them.