Move the code into a package (needed for PEP 561)
If we want to distribute type information the right way and for the
tools to pick it up we need to follow PEP 561 and include a file named
py.typed. Unfortunately there's no way to include the file with
module-only distributions[1]:

    This PEP does not support distributing typing information as part of
    module-only distributions. The code should be refactored into
    a package-based distribution and indicate that the package supports
    typing as described above.

Therefore we need to have a package first - this is what this commit
does.

[1] https://www.python.org/dev/peps/pep-0561/#packaging-type-information