Simplify pytest configuration
It seems to be doing fine on its own.

As a bonus this restores running doctests from the injector source code
which I broke when I made it a package (code in injector/__init__.py)
instead of a simple module. *But* we need to start skipping running
doctests for get_bindings on Python versions that don't support
Annotated (or don't support it well enough).