Update index.rst
Highlight that the code snippet will not work. Users most of the time will skim read the docs trying to spot snippets on how to use the code.
Since this is the only snippet on the welcome page, we might want to highlight that it's an example that won't work.